chore(electric-proxy): extract env types, enable observability, make secret optional
Move Env interface to dedicated env.ts file. Enable Cloudflare
observability. Make ELECTRIC_SECRET optional for local dev without it.
