+++ apps/electric-proxy/wrangler.jsonc
@@ -2,10 +2,13 @@
 	"name": "electric-proxy",
 	"account_id": "6ac34a81594ea52ff7ed2f24b9108e41",
 	"main": "src/index.ts",
-	"compatibility_date": "2025-01-01",
+	"compatibility_date": "2026-02-12",
 	"vars": {
 		"JWKS_URL": "https://api.superset.sh/api/auth/jwks",
 		"JWT_ISSUER": "https://api.superset.sh",
 		"JWT_AUDIENCE": "https://api.superset.sh"
+	},
+	"observability": {
+		"enabled": true
 	}
 }
